WordWise/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
pydantic
